Keep file paths relative
In order for file paths to stay relative to the gem root, the paths passed to
`ctags` cannot be absolute. From the `ctags(1)` man page:

> --tag-relative[=yes|no]
>      Indicates that the file paths recorded in the tag file should be relative
>      to the directory containing the tag file, rather than relative to the
>      current directory, unless the files supplied on the command line are
>      specified with absolute paths.

Simply adding `--tag-relative=yes` is not enough to make paths relative, but
it's still added to the command for the sake of being explicit about the desired
behavior.